requirements.txt
asgiref
async-timeout
attrs
autobahn
Automat
captcha
certifi
cffi
channels
channels-redis
charset-normalizer
constantly
cryptography
daphne
defusedxml
Django
django-debug-toolbar
django-ranged-response
django-simple-captcha
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
hyperlink
idna
incremental
msgpack
mysqlclient
oauthlib
Pillow
psycopg-binary
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
PyJWT
pyOpenSSL
python-dotenv
python3-openid
pytz
redis
requests
requests-oauthlib
service-identity
six
social-auth-app-django
social-auth-core
sqlparse
Twisted
twisted-iocpsupport
txaio
typing_extensions
tzdata
urllib3
zope.interface